plugin network: move to correct folder, moved old version, fixes and formatting
