+++ network/_pv_1_5_0/__init__.py
@@ -21,8 +21,11 @@
 
 import logging
 import socket
+import urllib.request
 import urllib.parse
-import lib.network
+import urllib.error
+#import lib.network to come soon, tidy up and document first
+import lib.connection
 
 from lib.model.smartplugin import SmartPlugin
 
@@ -34,162 +37,106 @@ NW_TCP_LISTEN   = 'nw_tcp_listen'
 NW_UDP_SEND     = 'nw_udp_send'
 
 
-class TCPDispatcher(lib.network.Tcp_server):
-    '''
-    Encapsulation class for using lib.network.Tcp_server class
-    '''
-    def __init__(self, parser, ip, port):
-        '''
-        Initializes the class
+class TCPHandler(lib.connection.Stream):
 
-        :param parser: callback function for handling data, called as parser(remoteip, 'tcp:<localip>:<localport>', '<data>')
-        :type parser: function
-        :param ip: local ip address to bind to
-        :type ip: str
-        :param port: local port to bind to
-        :type port: int
-        '''
-        super().__init__(port, ip, 1, b'\n')
+    def __init__(self, parser, dest, sock, source):
+        lib.connection.Stream.__init__(self, sock, source)
+        self.terminator = b'\n'
         self.parser = parser
-        self.dest = f'tcp:{ip}:{port}'
-        self.set_callbacks(data_received=self.handle_received_data, incoming_connection=self.handle_connection)
-        self.start()
+        self.dest = dest
+        self.source = source
 
-    def handle_connection(self, server, client):
-        '''
-        Handle incoming connection. Just used for debugging
-
-        :param server: Tcp_server object serving the connection
-        :type server: lib.network.Tcp_server
-        :param client: Client object for connection
-        :type client: lib.network.Client
-        '''
-        self.logger.debug(f'Incoming TCP connection from {client.name}')
-
-    def handle_received_data(self, server, client, data):
-        '''
-        Forward received data to parser callback
-
-        :param server: Tcp_server object serving the connection
-        :type server: lib.network.Tcp_server
-        :param client: Client object for connection
-        :type addr: lib.network.Client
-        :param data: received data
-        :type data: string
-        '''
-        self.logger.debug(f'Received packet from {client.ip}:{client.port} to {self.dest} with content "{data}"')
-        self.parser(client.ip, self.dest, data.strip())
+    def found_terminator(self, data):
+        self.parser(self.source, self.dest, data.decode(errors="ignore").strip())
+        self.close()
 
 
-class HTTPDispatcher(lib.network.Tcp_server):
-    '''
-    Encapsulation class for using lib.network.Tcp_server class with HTTP GET support
-    '''
+class TCPDispatcher(lib.connection.Server):
+
     def __init__(self, parser, ip, port):
-        '''
-        Initializes the class
-
-        :param parser: callback function for handling data, called as parser(remoteip, 'tcp:<localip>:<localport>', '<data>')
-        :type parser: function
-        :param ip: local ip address to bind to
-        :type ip: str
-        :param port: local port to bind to
-        :type port: int
-        '''
-        super().__init__(port, ip, 1, b'\n')
+        lib.connection.Server.__init__(self, ip, port)
         self.parser = parser
-        self.dest = f'http:{ip}:{port}'
-        self.set_callbacks(data_received=self.handle_received_data, incoming_connection=self.handle_connection)
-        self.start()
+        self.dest = "tcp:%s:%s" % (ip, port)
+        self.connect()
 
-    def handle_connection(self, server, client):
-        '''
-        Handle incoming connection. Just used for debugging
+    def handle_connection(self):
+        sock, address = self.accept()
+        if sock is None:
+            return
+        TCPHandler(self.parser, self.dest, sock, address)
 
-        :param server: Tcp_server object serving the connection
-        :type server: lib.network.Tcp_server
-        :param client: Client object for connection
-        :type client: lib.network.Client
-        '''
-        self.logger.debug(f'Incoming HTTP connection from {client.name}')
 
-    def handle_received_data(self, server, client, data):
-        '''
-        Forward received data to parser callback
+class HTTPHandler(lib.connection.Stream):
 
-        :param server: Tcp_server object serving the connection
-        :type server: lib.network.Tcp_server
-        :param client: Client object for connection
-        :type addr: lib.network.Client
-        :param data: received data
-        :type data: string
-        '''
-        self.logger.debug(f'Received packet from {client.ip}:{client.port} to {self.dest} via HTTP with content "{data}"')
+    def __init__(self, parser, dest, sock, source):
+        lib.connection.Stream.__init__(self, sock, source)
+        self.terminator = b"\r\n\r\n"
+        self.parser = parser
+        self.dest = dest
+        self.source = source
 
-        # find lines starting with GET and return remaining lines "HTTP-like" and return status
-        for line in data.splitlines():
+    def found_terminator(self, data):
+        for line in data.decode(errors="ignore").splitlines():
             if line.startswith('GET'):
                 request = line.split(' ')[1].strip('/')
-                if self.parser(client.ip, self.dest, urllib.parse.unquote(request)) is not False:
-                    client.send(b'HTTP/1.1 200 OK\r\n\r\n')
+                if self.parser(self.source, self.dest, urllib.parse.unquote(request)) is not False:
+                    self.send(b'HTTP/1.1 200 OK\r\n\r\n', close=True)
                 else:
-                    client.send(b'HTTP/1.1 400 Bad Request\r\n\r\n')
-                client.close()
+                    self.send(b'HTTP/1.1 400 Bad Request\r\n\r\n', close=True)
+                break
 
 
-class UDPDispatcher(lib.network.Udp_server):
-    '''
-    Encapsulation class for using lib.network.Udp_server class
-    '''
+class HTTPDispatcher(lib.connection.Server):
+
     def __init__(self, parser, ip, port):
-        '''
-        Initializes the class
-
-        :param parser: callback function for handling data, called as parser(remoteip, 'tcp:<localip>:<localport>', '<data>')
-        :type parser: function
-        :param ip: local ip address to bind to
-        :type ip: str
-        :param port: local port to bind to
-        :type port: int
-        '''
-        self.logger = logging.getLogger(__name__)
+        lib.connection.Server.__init__(self, ip, port)
         self.parser = parser
-        self.dest = f'udp:{ip}:{port}'
-        super().__init__(port, ip)
-        self.set_callbacks(data_received=self.handle_received_data)
-        self.start()
+        self.dest = "http:%s:%s" % (ip, port)
+        self.connect()
 
-    def handle_received_data(self, addr, data):
-        '''
-        Forward received data to parser callback
-
-        :param addr: address information about the remote client: ('<ip>', <port>)
-        :type addr: tuple
-        :param data: received data
-        :type data: string
-        '''
-        ip = addr[0]
-        addr = f'{addr[0]}:{addr[1]}'
-        self.logger.debug(f'{self.name}: incoming connection from {addr} to {self.dest}')
-        self.parser(ip, self.dest, data.strip())
+    def handle_connection(self):
+        sock, address = self.accept()
+        if sock is None:
+            return
+        HTTPHandler(self.parser, self.dest, sock, address)
 
 
-class Networkng(SmartPlugin):
-    '''
+class UDPDispatcher(lib.connection.Server):
+
+    def __init__(self, parser, ip, port):
+        self.logger = logging.getLogger(__name__)
+        lib.connection.Server.__init__(self, ip, port, proto='UDP')
+        self.dest = "udp:%s:%s" % (ip, port)
+        self.parser = parser
+        self.connect()
+
+    def handle_connection(self):
+        try:
+            data, addr = self.socket.recvfrom(4096)
+            ip = addr[0]
+            addr = "{}:{}".format(addr[0], addr[1])
+            self.logger.debug("{}: incoming connection from {} to {}".format(self._name, addr, self.address))
+        except Exception as e:
+            self.logger.exception("{}: {}".format(self._name, e))
+            return
+        self.parser(ip, self.dest, data.decode(errors="ignore").strip())
+
+
+class Network(SmartPlugin):
+    """
     Main class of the Plugin. Does all plugin specific stuff and provides
     the update functions for the items
-    '''
-    PLUGIN_VERSION = '1.6.0'
+    """
+    PLUGIN_VERSION = "1.4.1"
 
     generic_listeners = {}
     special_listeners = {}
     input_seperator = '|'
     socket_warning = 10
     socket_warning = 2
-    _dispatcher_list = []
 
-    def __init__(self, sh):
-        '''
+    def __init__(self, sh ):
+        """
         Initalizes the plugin.
 
         If you need the sh object at all, use the method self.get_sh() to get it. There should be almost no need for
@@ -199,7 +146,7 @@ class Networkng(SmartPlugin):
         use the SmartPlugin method get_parameter_value(parameter_name). Anywhere within the Plugin you can get
         the configured (and checked) value for a parameter by calling self.get_parameter_value(parameter_name). It
         returns the value in the datatype that is defined in the metadata.
-        '''
+        """
 
         # Call init code of parent class (SmartPlugin)
         super().__init__()
@@ -212,27 +159,24 @@ class Networkng(SmartPlugin):
         self.tcp_acl = self.parse_acl(self.get_parameter_value('tcp_acl'))
         self.udp_acl = self.parse_acl(self.get_parameter_value('udp_acl'))
         self.http_acl = self.parse_acl(self.get_parameter_value('http_acl'))
-
         if self.get_parameter_value('tcp') == 'yes':
             self.add_listener('tcp', self.get_parameter_value('ip'), self.get_parameter_value('port'), self.tcp_acl,
                               generic=True)
         if self.get_parameter_value('udp') == 'yes':
             self.add_listener('udp', self.get_parameter_value('ip'), self.get_parameter_value('port'), self.udp_acl,
                               generic=True)
-        if self.get_parameter_value('http') == 'yes':
-            self.logger.error('http parameter must be "no" or <port number>, but it is set to "yes". HTTP listener not enabled.')
-        elif self.get_parameter_value('http') != 'no':
+        if self.get_parameter_value('http') != 'no':
             self.add_listener('http', self.get_parameter_value('ip'), self.get_parameter_value('http'), self.http_acl,
                               generic=True)
 
     def udp(self, host, port, data):
-        '''
+        """
         This function writes given data to a host with specified port
 
         :param host: a host or an ip for destination
         :param port: a port number
         :param data: a string containing data
-        '''
+        """
         try:
             family, type, proto, canonname, sockaddr = socket.getaddrinfo(host, port)[0]
             sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
@@ -240,15 +184,14 @@ class Networkng(SmartPlugin):
             sock.close()
             del (sock)
         except Exception as e:
-            self.logger.warning(f'UDP: Problem sending data to {host}:{port}: {e}')
+            self.logger.warning("UDP: Problem sending data to {}:{}: {}".format(host, port, e))
             pass
         else:
-            self.logger.debug(f'UDP: Sending data to {host}:{port}: {data}')
+            self.logger.debug("UDP: Sending data to {}:{}: {}".format(host, port, data))
 
     def add_listener(self, proto, ip, port, acl='*', generic=False):
-        port = int(port)
-        dest = f'{proto}:{ip}:{port}'
-        self.logger.debug(f'Adding listener on: {dest}')
+        dest = "%s:%s:%s" % (proto, ip, port)
+        self.logger.debug("Adding listener on: {}".format(dest))
         if proto == 'tcp':
             dispatcher = TCPDispatcher(self.parse_input, ip, port)
         elif proto == 'udp':
@@ -258,10 +201,8 @@ class Networkng(SmartPlugin):
         else:
             return
 
-        if not dispatcher.listening():
+        if not dispatcher.connected:
             return False
-
-        self._dispatcher_list.append(dispatcher)
 
         acl = self.parse_acl(acl)
         if generic:
@@ -271,12 +212,12 @@ class Networkng(SmartPlugin):
         return True
 
     def parse_acl(self, acl):
-        '''
+        """
         Parse an acl which can contain either
             '*' which means accept data from all connection requests or
             a list of one or more ip which are allowed to connect
-        '''
-        self.logger.debug(f'parse_acl called with acl={acl}')
+        """
+        self.logger.debug("parse_acl called with acl='{}'".format(acl))
         if acl == ['*']:
             return False
         if isinstance(acl, str):
@@ -284,63 +225,61 @@ class Networkng(SmartPlugin):
         return acl
 
     def parse_input(self, source, dest, data):
-        '''
-        :param: source      # '<remoteip>:<port>' or '<remoteip>'
-        :param: dest        # '<proto>:<localip>:<port>'
+        """
+        :param: source
+        :param: dest
         :param: data either
-                        * `item|<item.path>|<value>`
-                        * `logic|<logic_name>|<value>`
-                        * `log|<loglevel>|message`  # loglevel could be info, warning or error
-        '''
-        self.logger.debug(f'parse_input called with source={source}, dest={dest} and data={data}')
-        proto = dest.split(':')[0].upper()
-        source_ip, __, __ = source.partition(':')
+                        * `item|item.path|value`
+                        * `logic|logic_name|value`
+                        * `log|loglevel|message` # loglevel could be info, warning or error
+        """
+        self.logger.debug("parse_input called with source={}, dest={} and data={}".format(source, dest, data))
         if dest in self.generic_listeners:
-            # decode data
-            data_list = data.split(self.input_seperator, 2)  # max 3 elements
-            if len(data_list) < 3:
-                self.logger.info(f'Ignoring input {data}. Format not recognized.')
+            inp = data.split(self.input_seperator, 2)  # max 3 elements
+            if len(inp) < 3:
+                self.logger.info("Ignoring input {}. Format not recognized.".format(data))
                 return False
-            typ, name, value = data_list
+            typ, name, value = inp
+            proto = dest.split(':')[0].upper()
+            source, __, port = source.partition(':')
             gacl = self.generic_listeners[dest]['acl']
 
             if typ == 'item':
                 if name not in self.generic_listeners[dest]['items']:
-                    self.logger.error(f'Item {name} not available in the generic listener.')
+                    self.logger.error("Item '{}' not available in the generic listener.".format(name))
                     return False
                 iacl = self.generic_listeners[dest]['items'][name]['acl']
                 if iacl:
-                    if source_ip not in iacl:
-                        self.logger.error(f'Item {name} acl doesn\'t permit updates from {source_ip}.')
+                    if source not in iacl:
+                        self.logger.error("Item '{}' acl doesn't permit updates from {}.".format(name, source))
                         return False
                 elif gacl:
-                    if source_ip not in gacl:
-                        self.logger.error(f'Generic network acl doesn\'t permit updates from {source_ip}.')
+                    if source not in gacl:
+                        self.logger.error("Generic network acl doesn't permit updates from {}.".format(source))
                         return False
                 item = self.generic_listeners[dest]['items'][name]['item']
-                self.logger.debug(f'Setting {item} to {value} (from {proto}:{source_ip})')
-                item(value, proto, source_ip)
+                item(value, proto, source)
 
             elif typ == 'logic':
                 if name not in self.generic_listeners[dest]['logics']:
-                    self.logger.error(f'Logic {name} not available in the generic listener.')
+                    self.logger.error("Logic '{}' not available in the generic listener.".format(name))
                     return False
                 lacl = self.generic_listeners[dest]['logics'][name]['acl']
                 if lacl:
-                    if source_ip not in lacl:
-                        self.logger.error(f'Logic {name} acl doesn\'t permit triggering from {source_ip}.')
+                    if source not in lacl:
+                        self.logger.error("Logic '{}' acl doesn't permit triggering from {}.".format(name, source))
                         return False
                 elif gacl:
-                    if source_ip not in gacl:
-                        self.logger.error(f'Generic network acl doesn\'t permit triggering from {source_ip}.')
+                    if source not in gacl:
+                        self.logger.error("Generic network acl doesn't permit triggering from {}.".format(source))
                         return False
                 logic = self.generic_listeners[dest]['logics'][name]['logic']
-                logic.trigger(proto, source_ip, value)
+                logic.trigger(proto, source, value)
 
             elif typ == 'log':
                 if gacl:
-                    if source_ip not in gacl:
-                        self.logger.error(f'Generic network acl doesn\'t permit log entries from {source_ip}')
+                    if source not in gacl:
+                        self.logger.error("Generic network acl doesn't permit log entries from {}".format(source))
                         return False
                 if name == 'info':
                     self.logger.info(value)
@@ -349,9 +288,9 @@ class Networkng(SmartPlugin):
                 elif name == 'error':
                     self.logger.error(value)
                 else:
-                    self.logger.warning(f'Unknown logging priority {name}. Data: {data}')
+                    self.logger.warning("Unknown logging priority '{}'. Data: '{}'".format(name, data))
             else:
-                self.logger.error(f'Unsupported key element {typ}. Data: {data}')
+                self.logger.error("Unsupporter key element {}. Data: {}".format(typ, data))
                 return False
 
         elif dest in self.special_listeners:
@@ -366,71 +305,66 @@ class Networkng(SmartPlugin):
                 lacl = self.special_listeners[dest]['logics'][entry]['acl']
                 logic = self.special_listeners[dest]['logics'][entry]['logic']
                 if lacl:
-                    if source_ip not in lacl:
-                        self.logger.error(f'Logic {logic.name} acl doesn\'t permit triggering from {source_ip}.')
+                    if source not in lacl:
+                        self.logger.error(
+                            "Logic '{}' acl doesn't permit triggering from {}.".format(logic.name, source))
                         return False
                 elif gacl:
-                    if source_ip not in gacl:
-                        self.logger.error(f'Generic network acl doesn\'t permit triggering from {source_ip}.')
+                    if source not in gacl:
+                        self.logger.error("Generic network acl doesn't permit triggering from {}.".format(source))
                         return False
-                logic.trigger('network', source_ip, data)
+                logic.trigger('network', source, data)
             for entry in self.special_listeners[dest]['items']:
                 lacl = self.special_listeners[dest]['items'][entry]['acl']
                 item = self.special_listeners[dest]['items'][entry]['item']
                 if lacl:
-                    if source_ip not in lacl:
-                        self.logger.error(f'Item {item.id()} acl doesn\'t permit triggering from {source_ip}.')
+                    if source not in lacl:
+                        self.logger.error("Item {0} acl doesn't permit triggering from {1}.".format(item.id(), source))
                         return False
                 elif gacl:
-                    if source_ip not in gacl:
-                        self.logger.error(f'Generic network acl doesn\'t permit triggering from {source_ip}.')
+                    if source not in gacl:
+                        self.logger.error("Generic network acl doesn't permit triggering from {0}.".format(source))
                         return False
-                self.logger.debug(f'Setting {item} to {data} (from {proto}:{source_ip})')
-                item(data, 'network', source_ip)
+                item(data, 'network', source)
         else:
-            self.logger.error(f'Destination {dest}, not in listeners!')
+            self.logger.error("Destination {}, not in listeners!".format(dest))
             return False
         return True
 
     def run(self):
-        '''
+        """
         Run method for the plugin
-        '''
-        self.logger.debug('Run method called')
+        """
+        self.logger.debug("Run method called")
         self.alive = True
 
     def stop(self):
-        '''
+        """
         Stop method for the plugin
-        '''
-        self.logger.debug('Stop method called')
-        for listener in self._dispatcher_list:
-            try:  # keep some ugly errors on keyboard interrupt hidden
-                listener.close()
-            except:
-                pass
-        self._dispatcher_list = []
+        """
+        self.logger.debug("Stop method called")
         self.alive = False
 
     def parse_logic(self, logic):
-        '''
+        """
         Called when the plugin is initialized to parse a logic
-        '''
+        """
         self.parse_obj(logic, 'logic')
 
     def parse_item(self, item):
-        '''
-        This is called when the plugin is initialized and
+        """
+        This is called when the plugin is initialized and 
         a param item is given to decide wether to act on it or not
-        '''
+        """
         self.parse_obj(item, 'item')
 
         if self.has_iattr(item.conf, NW_UDP_SEND):
-            self.logger.debug(f'parse item: {item}')
+            self.logger.debug("parse item: {}".format(item))
             return self.update_item
 
+
     def update_item(self, item, caller=None, source=None, dest=None):
-        '''
+        """
         Item has been updated
 
         This method is called, if the value of an item has been updated by SmartHomeNG.
@@ -441,15 +375,15 @@ class Networkng(SmartPlugin):
         :param caller: if given it represents the callers name
         :param source: if given it represents the source
         :param dest: if given it represents the dest
-        '''
+        """
         if self.alive and caller != self.get_shortname():
             # code to execute if the plugin is not stopped
             # and only, if the item has not been changed by this this plugin:
-            self.logger.info(f'Update item: {item.id()}, item has been changed outside this plugin')
+            self.logger.info("Update item: {}, item has been changed outside this plugin".format(item.id()))
 
             if self.has_iattr(item.conf, NW_UDP_SEND):
-                # nw_udp_send: '11.11.11.11:7777=command: itemvalue'    ## sends an UDP packet with 'command: ' and the current item value as payload
-                # instead of 'command' any text may be chosen. If somewhere right from equal sign
+                # nw_udp_send: "11.11.11.11:7777=command: itemvalue"    ## sends an UDP packet with 'command: ' and the current item value as payload
+                # instead of "command" any text may be chosen. If somewhere right from equal sign
                 # the word ``itemvalue``` appears, it will be replaced by the items value
                 addr, __, message = self.get_iattr_value(item.conf, NW_UDP_SEND).partition('=')
                 if not message:
@@ -460,11 +394,13 @@ class Networkng(SmartPlugin):
                 self.udp(host, port, message)
 
     def parse_obj(self, obj, obj_type):
-        '''
+        """
         parses either an item with item.conf or a logic with logic.conf
-        '''
+        """
         # NW_ACL, NW_UDP, NW_TCP
-        if obj_type in ('item', 'logic'):
+        if obj_type == 'item':
+            oid = obj.id()
+        elif obj_type == 'logic':
             oid = obj.id()
         else:
             return
@@ -484,12 +420,11 @@ class Networkng(SmartPlugin):
             if not ip:
                 ip = '0.0.0.0'
             dest = 'udp:' + ip + ':' + port
-            self.logger.debug(f'Adding listener {dest} for item {oid}')
             if dest not in self.special_listeners:
                 if self.add_listener('udp', ip, port):
                     self.special_listeners[dest][obj_type + 's'][oid] = {obj_type: obj, 'acl': acl}
                 else:
-                    self.logger.warning(f'Could not add listener {dest} for {oid}')
+                    self.logger.warning("Could not add listener {} for {}".format(dest, oid))
             else:
                 self.special_listeners[dest][obj_type + 's'][oid] = {obj_type: obj, 'acl': acl}
 
@@ -502,6 +437,6 @@ class Networkng(SmartPlugin):
                 if self.add_listener('tcp', ip, port):
                     self.special_listeners[dest][obj_type + 's'][oid] = {obj_type: obj, 'acl': acl}
                 else:
-                    self.logger.warning(f'Could not add listener {dest} for {oid}')
+                    self.logger.warning("Could not add listener {} for {}".format(dest, oid))
             else:
                 self.special_listeners[dest][obj_type + 's'][oid] = {obj_type: obj, 'acl': acl}
